:boom: [Machinist] BREAKING - lots of changes to the machines
Re-implemented machines to be redux-like. Still using mutable state, but all of the mutations are now described in a single place for each machine. Not sure how I feel about this right now, it adds a decent amount of boilerplate and may have increased the complexity of the code in a different way. It made running the createAnimationSequences on each phase simple though.